+++ src/containers/Flashcards/utils/index.js
@@ -1,0 +1,1 @@
+export { getRandomVocab } from './getRandomVocab';
