--- src/utils/getOnlyVocab.js
@@ -1,1 +1,0 @@
-export default (vocab) => vocab.replace(/\d:/gmi, '');
