+++ src/utils/vocabulary/getOnlyVocab.js
@@ -1,0 +1,1 @@
+export const getOnlyVocab = (vocab) => vocab.replace(/\d:/gmi, '');
