+++ src/utils/vocabulary/getVocabSpecificReading.js
@@ -1,6 +1,6 @@
 import { URL_SEPARATOR } from '@config/constants';
 
-export default (name) => {
+export const getVocabSpecificReading = (name) => {
   const newName = name.split(URL_SEPARATOR);
 
   return name.indexOf(URL_SEPARATOR) > 0 ? newName[2] : null;
