+++ src/utils/vocabulary/index.js
@@ -1,0 +1,2 @@
+export { getOnlyVocab } from './getOnlyVocab';
+export { getVocabSpecificReading } from './getVocabSpecificReading';
